+++ src/streamlink/session.py
@@ -44,6 +44,7 @@ class Streamlink:
             "hls-live-edge": 3,
             "hls-segment-attempts": 3,
             "hls-segment-ignore-names": [],
+            "hls-segment-include-names": [],
             "hls-segment-threads": 1,
             "hls-segment-timeout": 10.0,
             "hls-segment-stream-data": False,
@@ -119,6 +120,10 @@ class Streamlink:
 
         hls-segment-ignore-names (str[]) List of segment names without
                                  file endings which should get filtered out,
+                                 default: ``[]``
+
+        hls-segment-include-names (str[]) List of segment names with
+                                 file which should get filtered in,
                                  default: ``[]``
 
         hls-segment-threads      (int) The size of the thread pool used